+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/fragments/JsonFormFragment.java
@@ -18,6 +18,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
+import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
 import android.widget.ImageView;
@@ -46,6 +47,17 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import static com.vijay.jsonwizard.constants.JsonFormConstants.BOTTOM_NAVIGATION;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.BOTTOM_NAVIGATION_ORIENTATION;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.NEXT;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.NEXT_LABEL;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.NEXT_TYPE;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.PREVIOUS;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.PREVIOUS_LABEL;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.STEPNAME;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.SUBMIT;
+import static com.vijay.jsonwizard.constants.JsonFormConstants.SUBMIT_LABEL;
+
 /**
  * Created by vijay on 5/7/15.
  */
@@ -58,6 +70,13 @@ public class JsonFormFragment extends MvpFragment<JsonFormFragmentPresenter, Jso
     private JsonApi mJsonApi;
     private Map<String, List<View>> lookUpMap = new HashMap<>();
     public OnFieldsInvalid onFieldsInvalid;
+
+    private Button previousButton;
+    private Button nextButton;
+    private String stepName;
+    private LinearLayout bottomNavigation;
+
+    private BottomNavigationListener navigationListener;
 
     public static JsonFormFragment getFormFragment(String stepName) {
         JsonFormFragment jsonFormFragment = new JsonFormFragment();
@@ -90,20 +109,78 @@ public class JsonFormFragment extends MvpFragment<JsonFormFragmentPresenter, Jso
 
         mMainView = rootView.findViewById(R.id.main_layout);
         mScrollView = rootView.findViewById(R.id.scroll_view);
+        previousButton = rootView.findViewById(R.id.previous_button);
+        nextButton = rootView.findViewById(R.id.next_button);
+        bottomNavigation = rootView.findViewById(R.id.bottom_navigation_layout);
+        navigationListener = new BottomNavigationListener();
+        if (getArguments() != null) {
+            stepName = getArguments().getString(STEPNAME);
+        }
 
         setupToolbarBackButton();
+
+        JSONObject step = getStep(stepName);
+        if (step.optBoolean(BOTTOM_NAVIGATION)) {
+            initializeBottomNavigation(step, rootView);
+        }
         return rootView;
     }
 
     private void setupToolbarBackButton() {
         if (getArguments() != null) {
-            String stepName = getArguments().getString(JsonFormConstants.STEPNAME);
+            String stepName = getArguments().getString(STEPNAME);
             if (getStep(stepName).optBoolean(JsonFormConstants.DISPLAY_BACK_BUTTON)) {
                 getSupportActionBar().setHomeAsUpIndicator(getHomeUpIndicator());
                 setUpBackButton();
             }
         }
 
+    }
+
+    protected void initializeBottomNavigation(JSONObject step, View rootView) {
+        if (step.has(PREVIOUS)) {
+            previousButton.setVisibility(View.VISIBLE);
+            if (step.has(PREVIOUS_LABEL)) {
+                previousButton.setText(step.optString(PREVIOUS_LABEL));
+            }
+        }
+
+        if (step.has(NEXT)) {
+            nextButton.setVisibility(View.VISIBLE);
+            if (step.has(NEXT_LABEL)) {
+                nextButton.setText(step.optString(NEXT_LABEL));
+            }
+        }  else if (step.optString(NEXT_TYPE).equalsIgnoreCase(SUBMIT)) {
+            nextButton.setTag(R.id.submit, true);
+            nextButton.setVisibility(View.VISIBLE);
+            if (step.has(SUBMIT_LABEL)) {
+                nextButton.setText(step.optString(SUBMIT_LABEL));
+            } else {
+                nextButton.setText(R.string.submit);
+            }
+        } else if (!step.has(NEXT)) {
+            nextButton.setTag(R.id.submit, true);
+            nextButton.setVisibility(View.VISIBLE);
+            nextButton.setText(R.string.save);
+        }
+
+        if (step.has(BOTTOM_NAVIGATION_ORIENTATION)) {
+            // layout orientation
+            int orientation = "vertical".equals(step.optString(BOTTOM_NAVIGATION_ORIENTATION)) ? LinearLayout.VERTICAL : LinearLayout.HORIZONTAL;
+            bottomNavigation.setOrientation(orientation);
+            bottomNavigation.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT));
+            bottomNavigation.removeView(previousButton);
+            bottomNavigation.addView(previousButton);
+            // nav btn params
+            LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT);
+            int layoutMargin = Utils.pixelToDp((int) getContext().getResources().getDimension(R.dimen.bottom_navigation_margin), getContext());
+            params.setMargins(layoutMargin, layoutMargin, layoutMargin, layoutMargin);
+            previousButton.setLayoutParams(params);
+            nextButton.setLayoutParams(params);
+        }
+
+        rootView.findViewById(R.id.previous_button).setOnClickListener(navigationListener);
+        rootView.findViewById(R.id.next_button).setOnClickListener(navigationListener);
     }
 
     @DrawableRes
@@ -172,8 +249,7 @@ public class JsonFormFragment extends MvpFragment<JsonFormFragmentPresenter, Jso
 
     public boolean next() {
         try {
-            presenter.onNextClick(mMainView);
-            return true;
+            return presenter.onNextClick(mMainView);
         } catch (Exception e) {
             Log.e(TAG, Log.getStackTraceString(e));
         }
@@ -441,7 +517,7 @@ public class JsonFormFragment extends MvpFragment<JsonFormFragmentPresenter, Jso
         getActivity().getSupportFragmentManager().beginTransaction()
                 .setCustomAnimations(R.anim.enter_from_right, R.anim.exit_to_left, R.anim.enter_from_left,
                         R.anim.exit_to_right).replace(R.id.container, next).addToBackStack(next.getClass().getSimpleName())
-                .commit();
+                .commitAllowingStateLoss(); // use https://stackoverflow.com/a/10261449/9782187
     }
 
     public Menu getMenu() {
@@ -487,4 +563,22 @@ public class JsonFormFragment extends MvpFragment<JsonFormFragmentPresenter, Jso
     public boolean onMenuItemClick(MenuItem item) {
         return presenter.onMenuItemClick(item);
     }
+
+
+    protected class BottomNavigationListener implements View.OnClickListener {
+        @Override
+        public void onClick(View v) {
+            if (v.getId() == R.id.next_button) {
+                Object isSubmit = v.getTag(R.id.submit);
+                if (isSubmit != null && Boolean.valueOf(isSubmit.toString())) {
+                    save(false);
+                } else {
+                    next();
+                }
+            } else if (v.getId() == R.id.previous_button) {
+                getFragmentManager().popBackStack();
+            }
+        }
+    }
+
 }